Implement Access Denied for Unauthorized Users (#24). Close #24
This commit adds:

 - A new `AuthServiceBearerTokenAccessDeniedHandler` class that implements the `AccessDeniedHandler` interface to handle access denied exceptions for unauthorized users. The handler uses the `HandlerExceptionResolver` to resolve the exception and return a proper error message.

- A new `AuthServiceHandlerAdvice` class has been added to handle `AccessDeniedException` exceptions and return a 403 Forbidden response with a JSON error message.

- The `SecurityConfiguration` class has been updated to use the new `AuthServiceBearerTokenAccessDeniedHandler` as the access denied handler for the OAuth2 resource server.

- Controller testing implemented.

- Slight modification in test packages, and `com.seyed.ali.authenticationservice.exceptions.AuthServiceHandlerAdvice#handleAccessDeniedException(AccessDeniedException)`
