+++ src/main/java/com/seyed/ali/authenticationservice/exceptions/AuthServiceHandlerAdvice.java
@@ -1,6 +1,7 @@
 package com.seyed.ali.authenticationservice.exceptions;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -12,11 +13,11 @@ import static org.springframework.http.HttpStatus.FORBIDDEN;
 @RestControllerAdvice
 public class AuthServiceHandlerAdvice {
 
-    @ExceptionHandler({org.springframework.security.access.AccessDeniedException.class})
+    @ExceptionHandler({AccessDeniedException.class})
     @ResponseStatus(FORBIDDEN)
-    public ResponseEntity<?> handleAccessDeniedException(org.springframework.security.access.AccessDeniedException e) {
-        return ResponseEntity.ok(Map.of(
-                "code", FORBIDDEN,
+    public ResponseEntity<Map<String, String>> handleAccessDeniedException(AccessDeniedException e) {
+        return ResponseEntity.status(FORBIDDEN).body(Map.of(
+                "code", FORBIDDEN.toString(),
                 "message", "No permission.",
                 "server_message", e.getMessage()
         ));
